refactor: rpc.ts <-> request.ts
